requirement.txt
absl-py==1.4.0
accelerate==0.23.0
addict==2.4.0
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==6.0.0
cachetools==5.3.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
chicken==0.1.0
click==8.1.3
cmake==3.27.1
comm==0.1.3
contourpy==1.0.6
cycler==0.11.0
datasets==2.14.5
debugpy==1.6.7
decorator==5.1.1
deepface==0.0.79
defusedxml==0.7.1
dill==0.3.7
egg==0.2.0
exceptiongroup==1.0.4
executing==1.2.0
facenet-pytorch==2.5.3
fastjsonschema==2.16.3
filelock==3.12.2
fire==0.5.0
Flask==2.2.2
flatbuffers==23.5.26
fonttools==4.38.0
frozenlist==1.4.0
fsspec==2023.6.0
gast==0.4.0
gdown==4.7.1
gluonts==0.13.5
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.58.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.9.0
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.3
importlib-metadata==5.0.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
jstyleson==0.0.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keras==2.13.1
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.6
lit==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.5
mpmath==1.3.0
mtcnn==0.1.1
multidict==6.0.4
multiprocess==0.70.15
mysql-client==0.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.31
nbclassic==0.5.6
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==2.8.8
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omz-tools @ file:///home/veronica/repos/open_model_zoo/tools/model_tools
opencv-python==4.8.0.76
openvino==2023.0.2
openvino-dev==2023.0.2
openvino-telemetry==2023.1.1
opt-einsum==3.3.0
outcome==1.2.0
packaging==22.0
pandas==1.5.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.4.0
plotly==5.16.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
Pygments==2.15.1
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
regex==2023.8.8
requests==2.28.1
requests-oauthlib==1.3.1
retina-face==0.0.13
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
safetensors==0.3.3
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.9.3
seaborn==0.12.2
selenium==4.6.0
Send2Trash==1.8.2
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.6.2
sympy==1.12
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
terminado==0.17.1
texttable==1.6.7
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.7
tinycss2==1.2.1
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.33.2
trio==0.22.0
trio-websocket==0.9.2
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.12
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.7
wrapt==1.15.0
wsproto==1.2.0
xxhash==3.3.0
yarl==1.9.2
zipp==3.10.0
